pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gm-kit"
version = "0.1.0"
description = "Starter package for the GM Kit workspace."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "James Todd Garza" }]
dependencies = [
    "readchar",
    "typer",
    "rich",
    "PyMuPDF>=1.23.0",
    "tomli; python_version < \"3.11\"",
    "jsonschema>=4.0.0",
    "Pillow>=9.0.0",
]

[project.optional-dependencies]
dev = ["ruff", "pytest", "black", "isort", "mypy", "pytest-cov", "readchar", "typer", "bandit", "pymarkdownlnt"]
test = [
    "pytest",
    "pytest-cov",
    "pexpect; sys_platform != 'win32'",
    "wexpect; sys_platform == 'win32'",
]

[project.scripts]
gmkit = "gm_kit.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "PLR", "C90"]

[tool.ruff.lint.mccabe]
max-complexity = 20  # Grade E threshold (21+ will fail)

[tool.ruff.lint.pylint]
max-statements = 100  # Function length (approx 50-60 LLOC)

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E501", "PLR2004"]

[tool.hatch.build.targets.wheel]
packages = ["src/gm_kit"]
include = ["src/gm_kit/assets/**"]

[tool.hatch.build.targets.sdist]
include = ["src/gm_kit/assets/**"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["fitz"]
ignore_missing_imports = true
